refactor: change file memories to receive new function of authenticate by fastify jwt

+++ src/routes/memories.ts
@@ -1,14 +1,13 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '@/lib/prisma'
 import { z } from 'zod'
-import { request } from 'http'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.addHook('preHandler', async () => {
+  app.addHook('preHandler', async (request) => {
     await request.jwtVerify()
   })
 
-  app.post('/memories', async (request, reply) => {
+  app.post('/memories', async (request, _) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -22,14 +21,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        user_id: 'dasdasdasdasdas12312das=das',
+        user_id: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.get('/memories', async (request, reply) => {
+  app.get('/memories', async (request, _) => {
     const memories = await prisma.memory.findMany({
       where: {
         user_id: request.user.sub,
@@ -48,7 +47,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request, _) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string(),
     })
@@ -60,6 +59,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -78,7 +81,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params)
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -98,6 +111,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.user_id !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
